+++ packages/hooks/babel.config.js
@@ -1,0 +1,15 @@
+// eslint-disable-next-line no-undef
+module.exports = {
+  presets: [
+    [
+      '@babel/preset-env',
+      {
+        'targets': {
+          node: 'current',
+        },
+        'modules': 'commonjs',
+      },
+    ],
+    ['@babel/preset-react'],
+  ],
+};