+++ packages/hooks/src/index.js
@@ -1,0 +1,2 @@
+export * from './useEthscanBalance'
+export * from './useEthscanTokensBalance'